% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lls.Seurat.metabolism.R
\name{lls.Seurat.metabolism}
\alias{lls.Seurat.metabolism}
\title{Metabolic Pathway Activity Scoring}
\usage{
lls.Seurat.metabolism(
  sce,
  method = "VISION",
  missing_values = F,
  ncores = 2,
  metabolism.type = "LLS"
)
}
\arguments{
\item{sce}{A Seurat object containing RNA assay data.}

\item{method}{The method to use for scoring ("VISION", "AUCell", "ssGSEA", "gsva").}

\item{missing_values}{A boolean indicating whether to fill missing values.}

\item{ncores}{Number of cores to use for parallel processing.}

\item{metabolism.type}{Type of metabolism pathway to use ("KEGG", "REACTOME", "LLS").}
}
\value{
A Seurat object with metabolic pathway scores added.
}
\description{
This function performs metabolic pathway activity scoring using various methods.
}
